+++ src/plugins/ColumnListTransformer.ts
@@ -1,5 +1,5 @@
 import { NotionToMarkdown } from "notion-to-md";
-import { IPlugin, NotionBlock } from "../config/configuration";
+import { IPlugin, NotionBlock } from "./pluginTypes";
 
 export async function notionColumnListToMarkdown(
   notionToMarkdown: NotionToMarkdown,
